Add compiler for running powershell scripts
Add a compiler script to directly invoke powershell scripts and process
their output. Quickly jump through stack traces and see script output in
the quickfix.

Named this 'powershell' instead of ps1 since compilers tend to match
their command name and it's more identifiable than 'pwsh'.

In my quickfix, errors look like:

let g:ps1_efm_show_error_categories = 0

C:/script.ps1|11 col 5 error| The term 'Write-Ouput' is not recognized
as the name of a cmdlet, function, script file, or operable program.
Check the spelling of the  name, or if a path was included, verify that
the path is correct and try again.
|| +     Write-Ouput $content
|| +     ~~~~~~~~~~~

let g:ps1_efm_show_error_categories = 1

C:/script.ps1|11 col 5 error| The term 'Write-Ouput' is not recognized
as the name of a cmdlet, function, script file, or operable program.
Check the spelling of the  name, or if a path was included, verify that
the path is correct and try again.
|| +     Write-Ouput $content
|| +     ~~~~~~~~~~~
||     + CategoryInfo          : ObjectNotFound: (Write-Ouput:String) [], CommandNotFoundException
||     + FullyQualifiedErrorId : CommandNotFoundException